+++ R/utils.R
@@ -117,3 +117,7 @@ assert_subset <- function(x, y, type = "value") {
   }
   TRUE
 }
+
+SOMA_OBJECT_TYPE_METADATA_KEY <- "soma_object_type"
+SOMA_ENCODING_VERSION_METADATA_KEY <- "soma_encoding_version"
+SOMA_ENCODING_VERSION <- "0"